# backend/bookstore-contracts/target/generated-sources/openapi/src/main/java/com/project/bookstore/api/BooksApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.project.bookstore.api;

import com.project.bookstore.dto.Book;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-19T22:10:40.538194900+01:00[Europe/Brussels]")
@Validated
@Tag(name = "Books", description = "Book catalog operations")
public interface BooksApi {

    /**
     * POST /books : Create a new book
     *
     * @param book  (required)
     * @return Book created (status code 201)
     */
    @Operation(
        operationId = "createBook",
        summary = "Create a new book",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Book created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/books",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Book> createBook(
        @Parameter(name = "Book", description = "", required = true) @Valid @RequestBody Book book
    );


    /**
     * GET /books/{bookId} : Get a book by ID
     *
     * @param bookId  (required)
     * @return Book found (status code 200)
     *         or Book not found (status code 404)
     */
    @Operation(
        operationId = "getBookById",
        summary = "Get a book by ID",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Book found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            }),
            @ApiResponse(responseCode = "404", description = "Book not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/books/{bookId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Book> getBookById(
        @Parameter(name = "bookId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("bookId") String bookId
    );


    /**
     * GET /books : Get all books
     *
     * @return List of books (status code 200)
     */
    @Operation(
        operationId = "getBooks",
        summary = "Get all books",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of books", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Book.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/books",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<Book>> getBooks(
        
    );

}