# backend/bookstore-contracts/target/generated-sources/openapi/src/main/java/com/project/bookstore/api/OrdersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.project.bookstore.api;

import com.project.bookstore.dto.OrderRequest;
import com.project.bookstore.dto.OrderResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-19T22:10:40.538194900+01:00[Europe/Brussels]")
@Validated
@Tag(name = "Orders", description = "Order management operations")
public interface OrdersApi {

    /**
     * POST /orders : Create an order
     *
     * @param orderRequest  (required)
     * @return Order created (status code 201)
     */
    @Operation(
        operationId = "createOrder",
        summary = "Create an order",
        tags = { "Orders" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Order created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrderResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/orders",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<OrderResponse> createOrder(
        @Parameter(name = "OrderRequest", description = "", required = true) @Valid @RequestBody OrderRequest orderRequest
    );


    /**
     * GET /orders : Get all orders
     *
     * @return List of orders (status code 200)
     */
    @Operation(
        operationId = "getOrders",
        summary = "Get all orders",
        tags = { "Orders" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of orders", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OrderResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/orders",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<OrderResponse>> getOrders(
        
    );

}