# backend/bookstore-service/src/main/java/com/project/endpoints/dto/OrderItemDto.java
package com.project.endpoints.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class OrderItemDto {

    private String bookId;
    private int quantity;
}