# backend/bookstore-service/src/main/java/com/project/endpoints/dto/OrderRequestDto.java
package com.project.endpoints.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class OrderRequestDto {

    private List<OrderItemDto> items;
}